Create Flights_delay.py
The dashboard application consists of three components:

Title of the application
Component to enter input year
5 Charts conveying the different types of flight delay. 
The chart section is divided into three segments.
Carrier and Weather delay in the first segment.
National air system and Security delay in the second segment
Late aircraft delay in the third segment.

To Do List:
Design layout for the application.
Create a callback function. 
Add a callback decorator, define inputs and outputs.
Review the helper function that performs computation on the provided inputs.
Create 5 line graphs.
Run the application.